+++ src/features/Portal/Document/Header.tsx
@@ -8,6 +8,7 @@ import { documentService } from '@/services/document';
 import { useChatStore } from '@/store/chat';
 import { chatPortalSelectors } from '@/store/chat/selectors';
 import { oneLineEllipsis } from '@/styles';
+import { getDocumentRenderMode } from '@/utils/documentRenderMode';
 
 import AutoSaveHint from './AutoSaveHint';
 
@@ -20,6 +21,7 @@ const Header = () => {
   );
 
   const title = document?.filename || document?.title;
+  const isReadonly = !!document && getDocumentRenderMode(document).mode === 'highlight';
 
   if (!documentId) return null;
 
@@ -47,9 +49,11 @@ const Header = () => {
           {title}
         </Text>
       </Flexbox>
-      <Flexbox horizontal align={'center'} gap={8}>
-        <AutoSaveHint />
-      </Flexbox>
+      {!isReadonly && (
+        <Flexbox horizontal align={'center'} gap={8}>
+          <AutoSaveHint />
+        </Flexbox>
+      )}
     </Flexbox>
   );
 };
